Drug Combination Class

Initialize a new DrugCombination object

Add a new drug to the combination

Generate all possible combinations of drug dose levels

Update the combinations whenever there's a change in any of the Drug objects

Get all dose combinations

Get the number of dose levels

Get the level identifiers for dose combinations

Print a summary of the drug combination

Arguments

drugs

List of Drug objects

drug

A Drug object to add

combined

Logical indicating whether to return the total number of combined dose levels (TRUE) or a vector of dose levels for each drug (FALSE)

doseCombi_selected

Optional parameter to specify which dose combinations to return levels for

Value

List of dose combinations

Numeric value or vector of the number of dose levels

Vector of dose combination level identifiers

Details

A reference class that represents combinations of drugs. This class manages objects of the Drug-class class to create and manage combinations of different drugs at various dose levels.

Fields

drugs

List of Drug objects

doseCombinations

List of generated dose combinations

See also